aliennor-backend copy 2/requirements.txt
Django==2.0
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-tinymce==2.6.0
djangorestframework==3.7.3
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
PyJWT==1.5.3
pytz==2017.3